src/requirements.txt
flask
Flask-SQLAlchemy
Flask-JWT-Extended
requests
postgrest
psycopg2-binary